[DUOS-1904] Added uniform Actions component (#1710)
* fixed merge issue

* ???? Testing new actions component.

* updated unit tests?

* Fixed eslint

* fix admin console

* fix tests

* Fixing unit tests

* Apply suggestions from code review


* added unit test for Update button (used to be combined with the vote button)

* Update cypress/component/DarCollectionTable/admin_actions.spec.js


* fix codacy

* Fixed unit tests

* Fixed unit tests?

* merged with DUOS-1902, deleted unnecessary files
